refactor: extract UI utilities to _ui.py
Move StepTracker, get_key, select_with_arrows, BannerGroup, show_banner,
BANNER, and TAGLINE from __init__.py into the new _ui.py module. Re-export
all symbols from __init__.py to preserve the public API.
